requirements-dev.txt
arctic==1.74.0
numpy==1.14.3
pandas==0.24.2
dataclasses==0.6
nameko
luigi
neo4j
py2neo
pyyaml
pymongo
expandvars
PyGithub
grpcio

